fix: resolve BigIntForbidden panics for specta bindings
- Changed u64/usize → u32 in specta-annotated structs:
  SentinelConfig (max_local_events, batch_interval_secs)
  SafetyConfig (max_vm_lifetime_seconds)
  SandboxResult (duration_ms)
- Added u32→usize cast in egress.rs for VecDeque::with_capacity
- Verified: src/bindings.ts now generates 434 lines with 37 typed
  command wrappers and 20+ TypeScript interfaces

+++ src-tauri/src/orchestrator/egress.rs
@@ -32,7 +32,7 @@ impl EgressFilter {
     pub fn from_sam_logic(sam_logic: &SamLogic) -> Self {
         let mut allowlist = sam_logic.privacy.egress_allowlist.clone();
         allowlist.extend(sam_logic.privacy.user_allowlist.clone());
-        let max_events = sam_logic.privacy.sentinel.max_local_events;
+        let max_events = sam_logic.privacy.sentinel.max_local_events as usize;
 
         println!(
             "🛡️  Egress Filter initialized with {} allowed domains",
